+++ crates/fefix/src/prelude.rs
@@ -1,9 +1,11 @@
 //! Convenience re-exports of common traits and various items within `fefix`.
 
 pub use crate::dict::Dictionary;
-pub use crate::tagvalue::Configure;
+pub use crate::json::Configure as ConfigureJson;
+pub use crate::tagvalue::Configure as ConfigureTagValue;
 pub use crate::{
-    Buffer, FieldType, GetConfig, RandomFieldAccess, RepeatingGroup, SetField, TagU16,
+    Buffer, FieldType, GetConfig, RandomFieldAccess, RepeatingGroup, SetField, StreamingDecoder,
+    TagU16,
 };
 
 #[cfg(feature = "fix40")]