+++ crates/fefix/src/tagvalue/raw_decoder.rs
@@ -1,12 +1,17 @@
 use crate::tagvalue::{utils, Config, Configure, DecodeError};
-use crate::GetConfig;
+use crate::{Buffer, GetConfig, StreamingDecoder};
 use std::ops::Range;
 
 /// An immutable view over the contents of a FIX message by a [`RawDecoder`].
 #[derive(Debug)]
 pub struct RawFrame<T> {
+    /// Raw, untouched contents of the message. Includes everything from `BeginString <8>` up to
+    /// `CheckSum <8>`.
     pub data: T,
+    /// The range of bytes that address the value of `BeginString <8>`.
     pub begin_string: Range<usize>,
+    /// The range of bytes that address all contents after `MsgType <35>` and before `CheckSum
+    /// <10>`.
     pub payload: Range<usize>,
 }
 
@@ -100,13 +105,15 @@ where
         Self::default()
     }
 
-    /// Turns `self` into a [`RawDecoderBuffered`], allocating an internal
-    /// buffer.
-    pub fn buffered(self) -> RawDecoderBuffered<C> {
-        RawDecoderBuffered {
+    /// Adds a [`Buffer`] to `self`, turning it into a [`RawDecoderStreaming`].
+    pub fn streaming<B>(self, buffer: B) -> RawDecoderStreaming<B, C>
+    where
+        B: Buffer,
+    {
+        RawDecoderStreaming {
             config: self.config,
-            buffer: Vec::new(),
-            last_parser_state: ParserState::Empty,
+            buffer,
+            state: ParserState::Empty,
         }
     }
 
@@ -158,55 +165,42 @@ impl<C> GetConfig for RawDecoder<C> {
 enum ParserState {
     Empty,
     Header(HeaderInfo, usize),
-    Err(DecodeError),
+    Failed,
 }
 
 /// A [`RawDecoder`] that can buffer incoming data and read a stream of messages.
 #[derive(Debug)]
-pub struct RawDecoderBuffered<C = Config> {
+pub struct RawDecoderStreaming<B, C = Config> {
+    buffer: B,
     config: C,
-    buffer: Vec<u8>,
-    last_parser_state: ParserState,
+    state: ParserState,
 }
 
-impl<C> RawDecoderBuffered<C>
+impl<B, C> StreamingDecoder for RawDecoderStreaming<B, C>
 where
+    B: Buffer,
     C: Configure,
 {
-    /// Empties all contents of the internal buffer of `self`.
-    pub fn clear(&mut self) {
-        self.buffer.clear();
-        self.last_parser_state = ParserState::Empty;
+    type Buffer = B;
+    type Error = DecodeError;
+
+    fn buffer(&mut self) -> &mut B {
+        &mut self.buffer
     }
 
-    /// Provides a buffer that must be filled before re-attempting to deserialize
-    /// the next [`RawFrame`].
-    ///
-    /// # Panics
-    ///
-    /// Panics if the last call to [`RawDecoderBuffered::raw_frame`]
-    /// returned an [`Err`].
-    pub fn supply_buffer(&mut self) -> &mut [u8] {
-        match self.last_parser_state {
-            ParserState::Empty => {
-                // There's no point in validating a FIX message that is too short to
-                // ever be valid.
-                self.buffer.resize(utils::MIN_FIX_MESSAGE_LEN_IN_BYTES, 0);
-                self.buffer.as_mut_slice()
-            }
+    fn num_bytes_required(&self) -> usize {
+        match self.state {
+            ParserState::Empty => utils::MIN_FIX_MESSAGE_LEN_IN_BYTES,
             ParserState::Header(_, expected_len) => {
                 let old_len = self.buffer.as_slice().len();
-                self.buffer.resize(expected_len, 0);
-                &mut self.buffer.as_mut_slice()[old_len..]
+                expected_len - old_len
             }
-            ParserState::Err(_) => {
-                panic!("This decoder is not valid anymore and it shouldn't have been used.")
-            }
+            ParserState::Failed => 0,
         }
     }
 
-    pub fn parse(&mut self) {
-        match self.last_parser_state {
+    fn try_parse(&mut self) -> Result<Option<()>, Self::Error> {
+        match self.state {
             ParserState::Empty => {
                 let header_info =
                     HeaderInfo::parse(self.buffer.as_slice(), self.config().separator());
@@ -216,44 +210,43 @@ where
                         + header_info.nominal_body_len
                         + utils::FIELD_CHECKSUM_LEN_IN_BYTES;
 
-                    self.last_parser_state =
-                        ParserState::Header(header_info, expected_len_of_frame);
+                    self.state = ParserState::Header(header_info, expected_len_of_frame);
+                    Ok(None)
                 } else {
-                    self.last_parser_state = ParserState::Err(DecodeError::Invalid);
+                    Err(DecodeError::Invalid)
                 }
             }
-            ParserState::Header(_, _) => {}
-            ParserState::Err(_) => {}
-        }
-    }
-
-    /// Tries to deserialize the next [`RawFrame`] from the internal buffer. If
-    /// the internal buffer does not contain a complete message, returns an
-    /// [`Ok(None)`].
-    pub fn raw_frame<'a>(&'a self) -> Result<Option<RawFrame<&'a [u8]>>, DecodeError> {
-        match &self.last_parser_state {
-            ParserState::Empty => Ok(None),
-            ParserState::Err(e) => match e {
-                DecodeError::CheckSum => Err(DecodeError::CheckSum),
-                DecodeError::Invalid => Err(DecodeError::Invalid),
-                DecodeError::FieldPresence => Err(DecodeError::FieldPresence),
-                DecodeError::IO(_) => unreachable!("Can't have an I/O error here."),
-            },
-            ParserState::Header(header_info, _len) => {
-                let data = &self.buffer.as_slice();
-
-                Ok(Some(RawFrame {
-                    data,
-                    begin_string: header_info.field_0.clone(),
-                    payload: header_info.field_1.end + 1
-                        ..data.len() - utils::FIELD_CHECKSUM_LEN_IN_BYTES,
-                }))
-            }
+            ParserState::Header(_, _) => Ok(Some(())),
+            ParserState::Failed => panic!("Failed state"),
         }
     }
 }
 
-impl<C> GetConfig for RawDecoderBuffered<C> {
+impl<B, C> RawDecoderStreaming<B, C>
+where
+    B: Buffer,
+    C: Configure,
+{
+    /// Tries to deserialize the next [`RawFrame`] from the internal buffer. If
+    /// the internal buffer does not contain a complete message, returns an
+    /// [`Ok(None)`].
+    pub fn raw_frame(&self) -> RawFrame<&[u8]> {
+        if let ParserState::Header(header_info, _len) = &self.state {
+            let data = &self.buffer.as_slice();
+
+            RawFrame {
+                data,
+                begin_string: header_info.field_0.clone(),
+                payload: header_info.field_1.end + 1
+                    ..data.len() - utils::FIELD_CHECKSUM_LEN_IN_BYTES,
+            }
+        } else {
+            panic!("The message is not fully decoded. Check `try_parse` return value.");
+        }
+    }
+}
+
+impl<B, C> GetConfig for RawDecoderStreaming<B, C> {
     type Config = C;
 
     fn config(&self) -> &C {
@@ -377,13 +370,13 @@ mod test {
     }
 
     #[test]
-    fn new_buffered_decoder_has_no_current_frame() {
-        let decoder = new_decoder().buffered();
-        assert!(decoder.raw_frame().unwrap().is_none());
+    fn new_streaming_decoder_has_no_current_frame() {
+        let decoder = new_decoder().streaming(vec![]);
+        assert!(decoder.num_bytes_required() > 0);
     }
 
     #[test]
-    fn new_buffered_decoder() {
+    fn new_streaming_decoder() {
         let stream = {
             let mut stream = Vec::new();
             for _ in 0..42 {
@@ -394,15 +387,14 @@ mod test {
             stream
         };
         let mut i = 0;
-        let mut decoder = new_decoder().buffered();
-        let mut frame = None;
-        while frame.is_none() || i >= stream.len() {
-            let buf = decoder.supply_buffer();
+        let mut decoder = new_decoder().streaming(vec![]);
+        let mut ready = false;
+        while !ready || i >= stream.len() {
+            let buf = decoder.fillable();
             buf.clone_from_slice(&stream[i..i + buf.len()]);
             i += buf.len();
-            decoder.parse();
-            frame = decoder.raw_frame().unwrap();
+            ready = decoder.try_parse().unwrap().is_some();
         }
-        assert!(frame.is_some());
+        assert_eq!(decoder.raw_frame().begin_string(), b"FIX.4.2");
     }
 }