+++ examples/03_decode_fix_stream/src/main.rs
@@ -1,5 +1,5 @@
 use fefix::prelude::*;
-use fefix::tagvalue::{Config, Decoder};
+use fefix::tagvalue::Decoder;
 use std::io::{Cursor, Read};
 
 const FIX_MESSAGES: &[&[u8]] = &[
@@ -14,16 +14,16 @@ fn main() {
     let fix_dictionary = Dictionary::fix42();
     // Let's create a FIX decoder. This is an expensive operation, and it should
     // only be done once at the beginning of your program and/or FIX session.
-    let mut fix_decoder = Decoder::<Config>::new(fix_dictionary).buffered();
+    let mut fix_decoder = <Decoder>::new(fix_dictionary).streaming(vec![]);
     // In this case, the FIX message is specified using "|" rather than SOH
     // (ASCII 0x1) bytes. FerrumFIX supports this.
     fix_decoder.config_mut().set_separator(b'|');
     let mut stream = Cursor::new(fix_stream());
     loop {
-        let mut buffer = fix_decoder.supply_buffer();
+        let mut buffer = fix_decoder.fillable();
         // You *must* use `std::io::Read::read_exact`.
         stream.read_exact(&mut buffer).unwrap();
-        if let Ok(Some(())) = fix_decoder.parse() {
+        if let Ok(Some(())) = fix_decoder.try_parse() {
             let msg = fix_decoder.message();
             assert_eq!(msg.fv(fix42::BEGIN_STRING), Ok("FIX.4.2"));
         }